+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improved the styling of the checkboxes to consistently use the primary color in their states
+- Improved the account name display in the accounts table
 - Improved the name display in the activities table
 - Improved the user id display in the users table of the admin control panel
 - Improved the language localization for German (`de`)
@@ -18,7 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed a chart error on interaction by registering the annotation plugin early
 - Fixed an issue on the allocations page where clicking an account in the _By Account_ chart did not open the detail dialog
-- Fixed the SWR dropdown on the _FIRE_ page in dark mode
+- Restricted the maximum height of the import activities dialog
+- Fixed the dark mode styling of the safe withdrawal rate selector in the _FIRE_ section (experimental)
 
 ## 3.11.0 - 2026-06-14